pyproject.toml
[tool.poetry]
name = "mva"
version = "0.1.0"
description = ""
authors = ["valentynvovchak <valentyn.vovchak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
environs = "^9.5.0"
requests = "2.31.0"
uuid = "^1.30"
psycopg2-binary = "^2.9.6"
pillow = "^10.0.0"
django-cors-headers = "^4.1.0"
django-debug-toolbar = "^4.1.0"
django-extensions = "^3.2.3"
django-filter = "^23.2"
django-modeltranslation = "^0.18.10"
django-nested-admin = "^4.0.2"
django-nested-inline = "0.4.6"
djangorestframework = "^3.14.0"
markdown="3.4.4"
whitenoise = "^6.5.0"
django-json-widget = "^1.1.1"
django-reorder-admin = "^0.3.1"
setuptools = "^68.0.0"
gunicorn = "^21.2.0"
urllib3 = "^2.0.3"
uwsgi = "^2.0.21"
dj-database-url = "^2.0.0"
djangorestframework-api-key = "2.3.0"
drf-yasg = "^1.21.6"
python-dateutil = "^2.8.2"
django-extra-settings = "^0.10.0"
openpyxl = "^3.1.2"
drf-social-oauth2 = "2.1.3"
#pyjwt = "^2.6.0"
#django-allauth = "^0.51.0"
#dj-rest-auth = "^2.2.5"
#django-sslserver = "^0.22"
#djangorestframework-simplejwt = "^5.2.2"
#pytest = "^7.2.0"
#pytest-testdox = "^3.0.1"
#pytest-json-report = "^1.5.0"
#django-categories = "^1.9.2"
#elasticsearch = "<8.0.0"
#elasticsearch-dsl = "^7.4.1"
#django-elasticsearch-dsl = "^7.3"
celery = "^5.3.6"
django-celery-results = "^2.5.1"
redis = "^5.0.1"
pandas = "^2.2.0"
pyjwt = "2.9.0"
xhtml2pdf = "^0.2.16"
weasyprint = "^62.3"
pango = "^0.0.1"
python-http-client = ">=3.3.7,<3.4.0"
sendgrid = "6.9.7"
django-mptt = "^0.16.0"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
